evaluation stuff compiles on windows, fixed warnings from previous commits
git-svn-id: https://groke.otago.ac.nz/ant/ant@1513 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/directory_iterator_spam_filter.h
@@ -18,14 +18,14 @@ public:
 
 private:
 	ANT_directory_iterator *source;
-	long method;
+	long long method;
 	static char **docids;
 	static long long number_docs;
 
 	inline long should_index(char *docid);
 
 public:
-	ANT_directory_iterator_spam_filter(ANT_directory_iterator *source, char *filename, long long threshold, long method, long get_file = 0);
+	ANT_directory_iterator_spam_filter(ANT_directory_iterator *source, char *filename, long long threshold, long long method, long get_file = 0);
 	~ANT_directory_iterator_spam_filter();
 
 	virtual ANT_directory_iterator_object *first(ANT_directory_iterator_object *object);
